@@ -1,17 +1,5 @@
-import { DOCUMENT } from '@angular/common';
-import { Inject, Injectable } from '@angular/core';
-import { ReplaySubject, Observable } from 'rxjs';
-
-declare global {
-  interface Window {
-    MathJax: {
-      typesetPromise: () => void;
-      startup: {
-        promise: Promise<any>;
-      };
-    };
-  }
-}
+import { DOCUMENT, inject, Injectable } from '@angular/core';
+import { Observable, ReplaySubject } from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
@@ -19,11 +7,9 @@ declare global {
 export class MathService {
   private loaded$?: ReplaySubject<boolean>;
   private readonly source =
-    'https://cdn.jsdelivr.net/npm/mathjax@3.0.5/es5/mml-chtml.js';
-  private readonly integrity =
-    'sha256-CnzfCXjFj1REmPHgWvm/OQv8gFaxwbLKUi41yCU7N2s=';
+    'https://cdn.jsdelivr.net/npm/mathjax@4/tex-mml-chtml.js';
 
-  constructor(@Inject(DOCUMENT) private readonly document: Document) {}
+  private readonly document: Document = inject(DOCUMENT);
 
   registerMathJax(): Observable<boolean> {
     if (this.loaded$) {
@@ -35,7 +21,6 @@ export class MathService {
     const script = this.document.createElement('script');
     script.type = 'text/javascript';
     script.src = this.source;
-    script.integrity = this.integrity;
     script.async = true;
     script.crossOrigin = 'anonymous';
     script.onload = () => {
@@ -54,5 +39,18 @@ export class MathService {
       element.innerHTML = math;
       window.MathJax.typesetPromise();
     });
+  }
+}
+declare global {
+  interface Window {
+    MathJax: {
+      tex: {
+        inlineMath: { '[+]': [['$', '$']] };
+      };
+      typesetPromise: () => void;
+      startup: {
+        promise: Promise<any>;
+      };
+    };
   }
 }
