@@ -2,7 +2,6 @@ import {
   Directive,
   ElementRef,
   Input,
-  NgZone,
   OnChanges,
   OnInit,
   SimpleChanges,
@@ -11,17 +10,12 @@ import { MathService } from './math.service';
 
 @Directive({
   selector: '[appMath]',
-  standalone: false,
 })
 export class MathDirective implements OnInit, OnChanges {
   @Input() appMath = '';
   private readonly el: HTMLElement;
 
-  constructor(
-    private svc: MathService,
-    elementRef: ElementRef,
-    private ngZone: NgZone
-  ) {
+  constructor(private svc: MathService, elementRef: ElementRef) {
     this.el = elementRef.nativeElement;
   }
 
@@ -36,10 +30,8 @@ export class MathDirective implements OnInit, OnChanges {
   }
 
   private render() {
-    this.ngZone.runOutsideAngular(() => {
-      this.svc
-        .registerMathJax()
-        .subscribe(() => this.svc.render(this.el, this.appMath));
-    });
+    this.svc
+      .registerMathJax()
+      .subscribe(() => this.svc.render(this.el, this.appMath));
   }
 }
